backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
alembic==1.7.1
python-dotenv==0.19.0
psycopg2-binary==2.9.1
python-decouple==3.4
pydicom==2.2.2
pillow==8.3.2
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
torch==2.0.1
torchvision==0.15.2
pdf2image==1.16.3
pytest==6.2.5
pytest-asyncio==0.15.1
httpx==0.23.0 